Refactor CEvent and CEventEditableResponse
