+++ source/game_sa/Events/EventEditableResponse.cpp
@@ -22,39 +22,6 @@ void CEventEditableResponse::InjectHooks() {
     RH_ScopedInstall(ComputeResponseTaskOfType, 0x4B5730);
 }
 
-void CEventSpecial::InjectHooks() {
-    RH_ScopedClass(CEventSpecial);
-    RH_ScopedCategory("Events");
-
-    RH_ScopedInstall(Constructor, 0x4B1AE0);
-}
-
-void CEventFireNearby::InjectHooks() {
-    RH_ScopedClass(CEventFireNearby);
-    RH_ScopedCategory("Events");
-
-    RH_ScopedInstall(Constructor, 0x4B1F10);
-    RH_ScopedVirtualInstall(AffectsPed, 0x4B1F90);
-}
-
-void CEventDanger::InjectHooks() {
-    RH_ScopedClass(CEventDanger);
-    RH_ScopedCategory("Events");
-
-    RH_ScopedInstall(Constructor, 0x4B2600);
-    RH_ScopedVirtualInstall(AffectsPed, 0x4B5470);
-    RH_ScopedVirtualInstall(AffectsPedGroup, 0x4B54E0);
-    RH_ScopedVirtualInstall(GetSourceEntity, 0x4B2700);
-}
-
-void CEventSeenPanickedPed::InjectHooks() {
-    RH_ScopedClass(CEventSeenPanickedPed);
-    RH_ScopedCategory("Events");
-
-    RH_ScopedInstall(Constructor, 0x4B2080);
-    RH_ScopedVirtualInstall(AffectsPed, 0x4B53C0);
-}
-
 // 0x4AC450
 CEventEditableResponse::CEventEditableResponse() : CEvent() {
     m_bAddToEventGroup = true;
@@ -151,8 +118,7 @@ void CEventEditableResponse::InformRespectedFriends(CPed* ped) {
         }
 
         float fDmRadius = ped->GetIntelligence()->m_fDmRadius;
-        CVector direction = ped->GetPosition() - entity->GetPosition();
-        if (fDmRadius * fDmRadius > direction.SquaredMagnitude()) {
+        if (sq(fDmRadius) > DistanceBetweenPointsSquared(entity->GetPosition(), ped->GetPosition())) {
             auto clonedEvent = static_cast<CEventEditableResponse*>(Clone());
             clonedEvent->m_taskId = TASK_NONE;
             clonedEvent->m_bAddToEventGroup = false;
@@ -196,9 +162,9 @@ void CEventEditableResponse::ComputeResponseTaskType(CPed* ped, bool bDecisionMa
             GetEventType(),
             eventSourceType,
             ped->bInVehicle,
-            TASK_SIMPLE_INFORM_RESPECTED_FRIENDS, 
-            TASK_SIMPLE_INFORM_GROUP, 
-            TASK_SIMPLE_LOOK_AT_ENTITY_OR_COORD, 
+            TASK_SIMPLE_INFORM_RESPECTED_FRIENDS,
+            TASK_SIMPLE_INFORM_GROUP,
+            TASK_SIMPLE_LOOK_AT_ENTITY_OR_COORD,
             -1,
             bDecisionMakerTypeInGroup,
             m_taskId,
@@ -273,137 +239,4 @@ bool CEventEditableResponse::ComputeResponseTaskOfType(CPed* ped, int32 taskId) 
         unknownId
     );
     return taskId == outTaskId;
-}
-
-/*
- * CEventSpecial
- * */
-
-CEventSpecial* CEventSpecial::Constructor() {
-    this->CEventSpecial::CEventSpecial();
-    return this;
-}
-
-/*
- * CEventFireNearby
- * */
-
-// 0x4B1F10
-CEventFireNearby::CEventFireNearby(const CVector& position) : CEventEditableResponse() {
-    m_position = position;
-}
-
-CEventFireNearby* CEventFireNearby::Constructor(const CVector& position) {
-    this->CEventFireNearby::CEventFireNearby(position);
-    return this;
-}
-
-// 0x4B1F90
-bool CEventFireNearby::AffectsPed(CPed* ped) {
-    return CEventFireNearby::AffectsPed_Reversed(ped);
-}
-
-bool CEventFireNearby::AffectsPed_Reversed(CPed* ped)
-{
-    return !ped->GetTaskManager().Has<TASK_COMPLEX_EXTINGUISH_FIRES>() && ped->IsAlive();
-}
-
-/*
- * CEventDanger
- * */
-
-// 0x4B2600
-CEventDanger::CEventDanger(CEntity* dangerFrom, float dangerRadius) : CEventEditableResponse() {
-    m_dangerFrom = dangerFrom;
-    CEntity::SafeRegisterRef(m_dangerFrom);
-    m_dangerRadius = dangerRadius;
-}
-
-CEventDanger::~CEventDanger() {
-    CEntity::SafeCleanUpRef(m_dangerFrom);
-}
-
-CEventDanger* CEventDanger::Constructor(CEntity* dangerFrom, float dangerRadius) {
-    this->CEventDanger::CEventDanger(dangerFrom, dangerRadius);
-    return this;
-}
-
-// 0x4B5470
-bool CEventDanger::AffectsPed(CPed* ped) {
-    return CEventDanger::AffectsPed_Reversed(ped);
-}
-
-// 0x4B54E0
-bool CEventDanger::AffectsPedGroup(CPedGroup* pedGroup) {
-    return CEventDanger::AffectsPedGroup_Reversed(pedGroup);
-}
-
-// 0x4B2700
-CEntity* CEventDanger::GetSourceEntity() const {
-    return CEventDanger::GetSourceEntity_Reversed();
-}
-
-bool CEventDanger::AffectsPed_Reversed(CPed* ped) {
-    if (m_dangerFrom && m_dangerFrom != ped->m_pVehicle) {
-        CVector2D distance = ped->GetPosition() - m_dangerFrom->GetPosition();
-        if (m_dangerRadius * m_dangerRadius >= distance.SquaredMagnitude())
-            return ped->IsAlive();
-    }
-    return false;
-}
-
-bool CEventDanger::AffectsPedGroup_Reversed(CPedGroup* pedGroup) {
-    if (GetSourceEntity() && GetSourceEntity()->IsPed()) {
-        CPed* leader = pedGroup->GetMembership().GetLeader();
-        if (leader) {
-            CVector2D distance = leader->GetPosition() - m_dangerFrom->GetPosition();
-            return distance.SquaredMagnitude() <= m_dangerRadius * m_dangerRadius;
-        }
-    }
-    return false;
-}
-
-CEntity* CEventDanger::GetSourceEntity_Reversed() const {
-    if (m_dangerFrom && !m_dangerFrom->IsPed() && m_dangerFrom->IsVehicle()) {
-        CVehicle* vehicle = m_dangerFrom->AsVehicle();
-        if (vehicle->m_pDriver)
-            return vehicle->m_pDriver;
-    }
-    return m_dangerFrom;
-}
-
-/*
- * CEventSeenPanickedPed
- * */
-
-// 0x4B2080
-CEventSeenPanickedPed::CEventSeenPanickedPed(CPed* ped) : CEventEditableResponse() {
-    m_ped = ped;
-    CEntity::SafeRegisterRef(m_ped);
-}
-
-// 0x4B2130
-CEventSeenPanickedPed::~CEventSeenPanickedPed() {
-    CEntity::SafeCleanUpRef(m_ped);
-}
-
-CEventSeenPanickedPed* CEventSeenPanickedPed::Constructor(CPed* ped) {
-    this->CEventSeenPanickedPed::CEventSeenPanickedPed(ped);
-    return this;
-}
-
-// 0x4B53C0
-bool CEventSeenPanickedPed::AffectsPed(CPed* ped) {
-    return CEventSeenPanickedPed::AffectsPed_Reversed(ped);
-}
-
-bool CEventSeenPanickedPed::AffectsPed_Reversed(CPed* ped) {
-    if (!ped->IsPlayer() && m_ped && m_ped != ped) {
-        CEvent* currentEvent = m_ped->GetEventHandlerHistory().GetCurrentEvent();
-        if (currentEvent && currentEvent->GetSourceEntity()) {
-            CVector distance = ped->GetPosition() - m_ped->GetPosition();
-            return distance.SquaredMagnitude() < 100.0f;
-        }
-    }
-    return false;
 }
