+++ source/game_sa/Events/EventGunShot.cpp
@@ -4,8 +4,7 @@
 
 float& CEventGunShot::ms_fGunShotSenseRangeForRiot2 = *(float*)0x8A625C;
 
-void CEventGunShot::InjectHooks()
-{
+void CEventGunShot::InjectHooks() {
     RH_ScopedClass(CEventGunShot);
     RH_ScopedCategory("Events");
 
@@ -16,9 +15,14 @@ void CEventGunShot::InjectHooks()
     RH_ScopedVirtualInstall(CloneEditable, 0x4B6B20);
 }
 
+CEventGunShot* CEventGunShot::Constructor(CEntity* entity, CVector startPoint, CVector endPoint, bool bHasNoSound) { this->CEventGunShot::CEventGunShot(entity, startPoint, endPoint, bHasNoSound); return this; }
+bool CEventGunShot::AffectsPed(CPed* ped) { return CEventGunShot::AffectsPed_Reversed(ped); }
+bool CEventGunShot::IsCriminalEvent() { return CEventGunShot::IsCriminalEvent_Reversed(); }
+bool CEventGunShot::TakesPriorityOver(const CEvent& refEvent) { return CEventGunShot::TakesPriorityOver_Reversed(refEvent); }
+CEventEditableResponse* CEventGunShot::CloneEditable() { return CEventGunShot::CloneEditable_Reversed(); }
+
 // 0x4AC610
-CEventGunShot::CEventGunShot(CEntity* entity, CVector startPoint, CVector endPoint, bool bHasNoSound)
-{
+CEventGunShot::CEventGunShot(CEntity* entity, CVector startPoint, CVector endPoint, bool bHasNoSound) : CEventEditableResponse() {
     m_startPoint = startPoint;
     m_endPoint = endPoint;
     m_entity = entity;
@@ -26,43 +30,12 @@ CEventGunShot::CEventGunShot(CEntity* entity, CVector startPoint, CVector endPoi
     CEntity::SafeRegisterRef(m_entity);
 }
 
-CEventGunShot::~CEventGunShot()
-{
+CEventGunShot::~CEventGunShot() {
     CEntity::SafeCleanUpRef(m_entity);
 }
 
-CEventGunShot* CEventGunShot::Constructor(CEntity* entity, CVector startPoint, CVector endPoint, bool bHasNoSound)
-{
-    this->CEventGunShot::CEventGunShot(entity, startPoint, endPoint, bHasNoSound);
-    return this;
-}
-
 // 0x4B2CD0
-bool CEventGunShot::AffectsPed(CPed* ped)
-{
-    return CEventGunShot::AffectsPed_Reversed(ped);
-}
-
-// 0x4AC810
-bool CEventGunShot::IsCriminalEvent()
-{
-    return CEventGunShot::IsCriminalEvent_Reversed();
-}
-
-// 0x4AC780
-bool CEventGunShot::TakesPriorityOver(const CEvent& refEvent)
-{
-    return CEventGunShot::TakesPriorityOver_Reversed(refEvent);
-}
-
-// 0x4B6B20
-CEventEditableResponse* CEventGunShot::CloneEditable()
-{
-    return CEventGunShot::CloneEditable_Reversed();
-}
-
-bool CEventGunShot::AffectsPed_Reversed(CPed* ped)
-{
+bool CEventGunShot::AffectsPed_Reversed(CPed* ped) {
     if (!m_entity)
         return false;
 
@@ -81,20 +54,20 @@ bool CEventGunShot::AffectsPed_Reversed(CPed* ped)
         if (ped->IsAlive() && ped != m_entity && !ped->IsPlayer()) {
             float fGunShotRange = 45.0f;
             if (CEventGunShot::ms_fGunShotSenseRangeForRiot2 <= 0.0f) {
-                if (ped->m_nCreatedBy == PED_MISSION)
+                if (ped->IsCreatedByMission()) {
                     fGunShotRange = ped->GetIntelligence()->GetPedFOVRange();
-            }
-            else if (ped->m_nCreatedBy == PED_MISSION) {
+                }
+            } else if (ped->IsCreatedByMission()) {
                 fGunShotRange = CEventGunShot::ms_fGunShotSenseRangeForRiot2;
             }
-            CVector distance = ped->GetPosition() - m_entity->GetPosition();
-            if (distance .SquaredMagnitude() <= fGunShotRange * fGunShotRange) {
+
+            const auto& pedPos = ped->GetPosition();
+            if (DistanceBetweenPointsSquared(m_entity->GetPosition(), pedPos) <= sq(fGunShotRange)) {
                 if (!m_bHasNoSound)
                     return true;
 
-                CVector dist = m_startPoint - ped->GetPosition();
-                if (DotProduct(dist, ped->GetForward()) > 0.0f) {
-                    if (CWorld::GetIsLineOfSightClear(m_startPoint, ped->GetPosition(), true, true, true, true, true, false, false))
+                if (DotProduct(m_startPoint - pedPos, ped->GetForward()) > 0.0f) {
+                    if (CWorld::GetIsLineOfSightClear(m_startPoint, pedPos, true, true, true, true, true, false, false))
                         return true;
                 }
             }
@@ -103,13 +76,13 @@ bool CEventGunShot::AffectsPed_Reversed(CPed* ped)
     return false;
 }
 
-bool CEventGunShot::IsCriminalEvent_Reversed()
-{
+// 0x4AC810
+bool CEventGunShot::IsCriminalEvent_Reversed() {
     return m_entity && m_entity->IsPed() && m_entity->AsPed()->IsPlayer();
 }
 
-bool CEventGunShot::TakesPriorityOver_Reversed(const CEvent& refEvent)
-{
+// 0x4AC780
+bool CEventGunShot::TakesPriorityOver_Reversed(const CEvent& refEvent) {
     if (refEvent.GetEventType() == GetEventType()) {
         bool bIsPlayer = false;
         bool otherPedIsPlayer = false;
@@ -127,7 +100,7 @@ bool CEventGunShot::TakesPriorityOver_Reversed(const CEvent& refEvent)
     return GetEventPriority() >= refEvent.GetEventPriority();
 }
 
-CEventEditableResponse* CEventGunShot::CloneEditable_Reversed()
-{
+// 0x4B6B20
+CEventEditableResponse* CEventGunShot::CloneEditable_Reversed() {
     return new CEventGunShot(m_entity, m_startPoint, m_endPoint, m_bHasNoSound);
 }
