+++ source/game_sa/Events/EventPassObject.h
@@ -1,0 +1,31 @@
+#pragma once
+
+#include "Event.h"
+
+class CEventPassObject : public CEvent {
+public:
+    CEntity* m_giver;
+    bool     m_dontPassObject;
+
+public:
+    CEventPassObject(CEntity* giver, bool dontPassObject);
+    ~CEventPassObject() override;
+
+    eEventType GetEventType() const override { return EVENT_PASS_OBJECT; }
+    int32 GetEventPriority() const override { return 11; }
+    int32 GetLifeTime() override { return 0; }
+    CEventPassObject* Clone() override { return new CEventPassObject(m_giver, m_dontPassObject); }
+    bool AffectsPed(CPed* ped) override { return true; }
+    bool IsValid(CPed* ped) override;
+
+private:
+    friend void InjectHooksMain();
+    static void InjectHooks();
+
+    CEventPassObject* Constructor(CEntity* giver, bool dontPassObject);
+    CEventPassObject* Destructor();
+
+    bool IsValid_Reversed(CPed* ped);
+};
+
+VALIDATE_SIZE(CEventPassObject, 0x14);
