+++ source/game_sa/Events/EventRevived.cpp
@@ -1,0 +1,8 @@
+#include "StdInc.h"
+
+#include "EventRevived.h"
+
+// 0x4AECB0
+bool CEventRevived::AffectsPed(CPed* ped) {
+    return !ped->IsCreatedByMission() && !ped->IsAlive();
+}
