+++ source/game_sa/Events/EventSpecial.cpp
@@ -1,0 +1,3 @@
+#include "StdInc.h"
+
+#include "EventSpecial.h"
